003_process_videos/fast_graber_frame/requirements.txt
opencv-python
opencv-python-headless
ffmpeg-python
tqdm
pyyaml
pyopencl  # для AMD OpenCL